+++ CHANGELOG.md
@@ -2,11 +2,15 @@
 
 ## Upcoming version 0.27.0 (????-??-??)
 
-This release has been tested with PHP 8.5.
+PHP 8.2 or newer is now required! This release has been tested with PHP 8.5.
 
 ### Added
 
 * Plugin has been tested with PHP 8.5 [#171](https://github.com/chesio/bc-security/issues/171).
+
+### Changed
+
+* PHP 8.2 is required [#165](https://github.com/chesio/bc-security/issues/165).
 
 ## Version 0.26.0 (2025-04-17)